refactor(tools): split dataset.py into focused submodules
Closes #4. tools/dataset.py was 1745 lines / 17 tools -- past the
comfortable inline-grep threshold the engineer persona review flagged
during v1.0 ("when I'm asked to add download_dataset_bag, I'm scrolling
through 1700 lines of unrelated tools to find my anchor"). Split into
a package mirroring the read/mutate/complex grouping that was already
visible in the source's declaration order:

    tools/dataset/__init__.py   # register aggregator + helper re-exports
    tools/dataset/read.py       # 7 read tools + 4 shared helpers
    tools/dataset/mutate.py     # 7 simple mutation tools
    tools/dataset/complex.py    # 3 complex tools (cache, denormalize, split)

Pure refactor: zero behavior change, zero LLM-visible surface change,
zero test change. The 4 helpers (_summarize_dataset and 3 _*_impl
functions consumed by resources/ml.py and resources/rag.py) live in
read.py and are re-exported from __init__.py so all existing import
paths -- from deriva_ml_mcp.tools import dataset as dataset_module
(used by tests/test_dataset.py) and from deriva_ml_mcp.tools.dataset
import _list_datasets_impl (used by both resource modules) -- continue
to resolve identically.

The four names that test code patches as a single canonical site --
audit_event, get_ml, Dataset, _split_dataset -- are imported at the
package __init__.py level and accessed from the read/mutate/complex
submodules via attribute lookup on the parent package
(_pkg.audit_event, _pkg.get_ml, etc.) rather than per-submodule
``from ... import`` bindings. This preserves the pre-split single-file
patch surface: a single patch("deriva_ml_mcp.tools.dataset.<name>")
still redirects every call across all three submodules in one shot.

The original tools/dataset.py is deleted (no backwards-compat shim --
workspace convention). The package's __init__.py is the new public
face.

docs/coverage.md updated: each of the 17 dataset tool rows in the
Tools table now points at its specific submodule (read.py / mutate.py
/ complex.py) under tools/dataset/. The dataset-element-types row in
the Resources backfill section was updated to point at read.py too,
since the tool now lives there.

Tests: 225 unit + 14 integration -> 225 unit + 14 integration. No
test files modified. tests/test_plugin.py's _DATASET_TOOLS frozenset
still pins all 17 tool names; tests/test_dataset.py's existing
import pattern works unchanged.

ruff check + format clean.
